More documentation and cleanup.
